refactors bounties into new nav structure, tests successful, removed redundant navigation graphs
