--- app/src/main/res/navigation/request_navigation.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:id="@+id/request_navigation"
-        app:startDestination="@+id/navigation_request">
-
-    <fragment
-            android:id="@+id/navigation_request"
-            android:name="com.hover.stax.requests.NewRequestFragment"
-            android:label="@string/nav_request"
-            tools:layout="@layout/fragment_request">
-        <action
-                android:id="@+id/action_navigation_request_to_accountsFrag"
-                app:destination="@id/accountsFrag" />
-    </fragment>
-
-    <fragment
-            android:id="@+id/accountsFrag"
-            android:name="com.hover.stax.account.AccountsFragment"
-            android:label="AccountsFragment"
-            tools:layout="@layout/fragment_accounts">
-        <action
-                android:id="@+id/action_accountsFrag_to_navigation_request"
-                app:destination="@id/navigation_request"
-                app:popUpTo="@id/navigation_request"
-                app:popUpToInclusive="true" />
-    </fragment>
-</navigation>
