--- app/src/main/res/navigation/transfer_navigation.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:id="@+id/transfer_navigation"
-        app:startDestination="@+id/navigation_transfer">
-
-    <fragment
-            android:id="@+id/navigation_transfer"
-            android:name="com.hover.stax.transfers.TransferFragment"
-            android:label="@string/nav_transfer"
-            tools:layout="@layout/fragment_transfer">
-        <action
-                android:id="@+id/action_navigation_transfer_to_accountsFragment"
-                app:destination="@id/accountsFragment" />
-    </fragment>
-    <fragment
-            android:id="@+id/accountsFragment"
-            android:name="com.hover.stax.account.AccountsFragment"
-            android:label="AccountsFragment"
-            tools:layout="@layout/fragment_accounts">
-        <action
-                android:id="@+id/action_accountsFragment_to_navigation_transfer"
-                app:destination="@id/navigation_transfer"
-                app:popUpTo="@id/navigation_transfer"
-                app:popUpToInclusive="true" />
-    </fragment>
-
-
-</navigation>
